Worker refactoring
