+++ worker/src/lib.rs
@@ -1,0 +1,3 @@
+pub mod config;
+pub mod heartbeat;
+pub mod registration;
